Removed multiple implementations of ParallelExecutionEngine.
