+++ Expressions/Expressions/Expressions.csproj
@@ -66,10 +66,7 @@
     <Compile Include="Execution\ExecutionEngine.cs" />
     <Compile Include="Execution\IChangeListener.cs" />
     <Compile Include="Execution\NotificationResultCollection.cs" />
-    <Compile Include="Execution\Parallel\ParallelArrayExecutionEngine.cs" />
-    <Compile Include="Execution\Parallel\ParallelExecutionEngine.cs" />
-    <Compile Include="Execution\Parallel\ParallelForeachExecutionEngine.cs" />
-    <Compile Include="Execution\Parallel\ParallelQueueExecutionEngine.cs" />
+    <Compile Include="Execution\ParallelExecutionEngine.cs" />
     <Compile Include="Execution\PropertyChangeListener.cs" />
     <Compile Include="Execution\SequentialExecutionEngine.cs" />
     <Compile Include="ExpressionHelper.cs" />
