+++ NMF.sln
@@ -128,12 +128,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Incerator", "Tools\Incerato
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Analysis", "Analysis\Analysis\Analysis.csproj", "{09761C44-D1F2-4DB7-B33C-FCDBF260388C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrainBenchmark", "Expressions\Tests\TrainBenchmark\TrainBenchmark.csproj", "{32068384-1495-4002-9F26-CCFF63405B92}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DgmlVisualizer", "Expressions\Tests\DgmlVisualizer\DgmlVisualizer.csproj", "{AEB08821-4040-4F7E-8367-E9605125A277}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClassDiagramOptimization", "Expressions\Tests\TTC2016CRA\ClassDiagramOptimization.csproj", "{88851106-A711-4CD8-B2D8-5DBD7A42156E}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -316,18 +310,6 @@ Global
 		{09761C44-D1F2-4DB7-B33C-FCDBF260388C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{09761C44-D1F2-4DB7-B33C-FCDBF260388C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{09761C44-D1F2-4DB7-B33C-FCDBF260388C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{32068384-1495-4002-9F26-CCFF63405B92}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{32068384-1495-4002-9F26-CCFF63405B92}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{32068384-1495-4002-9F26-CCFF63405B92}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{32068384-1495-4002-9F26-CCFF63405B92}.Release|Any CPU.Build.0 = Release|Any CPU
-		{AEB08821-4040-4F7E-8367-E9605125A277}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{AEB08821-4040-4F7E-8367-E9605125A277}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{AEB08821-4040-4F7E-8367-E9605125A277}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{AEB08821-4040-4F7E-8367-E9605125A277}.Release|Any CPU.Build.0 = Release|Any CPU
-		{88851106-A711-4CD8-B2D8-5DBD7A42156E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{88851106-A711-4CD8-B2D8-5DBD7A42156E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{88851106-A711-4CD8-B2D8-5DBD7A42156E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{88851106-A711-4CD8-B2D8-5DBD7A42156E}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -383,8 +365,5 @@ Global
 		{1F67AE40-F8B5-4FD0-9560-ACF1EA6434B4} = {F71EE711-DD09-4ECB-B060-1A737D573AF3}
 		{9ACD49E1-4EE2-4CB0-832F-514D7E475693} = {C933F88A-DC6F-4D15-9816-6D47EC964AC4}
 		{09761C44-D1F2-4DB7-B33C-FCDBF260388C} = {DB96042F-8815-430D-8345-362F4266F0FE}
-		{32068384-1495-4002-9F26-CCFF63405B92} = {0B9F756F-763B-4BEB-A592-78DBD370B3A9}
-		{AEB08821-4040-4F7E-8367-E9605125A277} = {0B9F756F-763B-4BEB-A592-78DBD370B3A9}
-		{88851106-A711-4CD8-B2D8-5DBD7A42156E} = {0B9F756F-763B-4BEB-A592-78DBD370B3A9}
 	EndGlobalSection
 EndGlobal
